chat/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=0.19.0
tenacity>=8.0.0
requests>=2.28.0
aiohttp>=3.8.0
PyPDF2
python-docx
python-multipart>=0.0.5
loguru==0.7.2
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.1.1
cryptography
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib==1.7.4
bcrypt==4.0.1
email-validator>=1.1.3
redis>=5.0.0
asyncpg>=0.29.0
langgraph==0.3.25
langchain-deepseek==0.1.3
langchain_neo4j==0.4.0
langchain_community==0.3.21
langgraph-checkpoint-sqlite==2.0.6

# Wechat MP Server Dependencies
Flask>=3.0.0
wx-mp-svr
wx-crypt

# GraphRAG Core Dependencies（排除 Azure 存储后端，使用 LanceDB）
graphrag==2.7.1
environs>=11.0.0
lancedb>=0.17.0
aiofiles>=24.1.0
fnllm>=0.2.3
json-repair>=0.30.3
nltk>=3.9.1
tiktoken>=0.8.0
numpy>=1.25.2,<2.0.0
networkx>=3.4.2
pandas>=2.2.3
pyarrow>=15.0.0
pyyaml>=6.0.2
rich>=13.9.4
devtools>=0.12.2
future>=1.0.0
typer>=0.15.1
tqdm>=4.67.1
gensim>=4.3.3
graspologic>=3.4.1

# 阿里云 OSS（AI 生成图片上传到云存储）
oss2>=2.18.0

# SILK 音频解码（企业微信语音消息 .silk 格式解码为 PCM）
pilk>=0.2.0

# 语音识别 ASR 依赖
# sherpa-onnx: 本地 SenseVoice Small 语音识别引擎（需要提前下载模型）
sherpa-onnx>=1.10.0
# numpy 需要兼容版本
numpy>=1.25.2,<2.0.0
